Drastic performance improvement
Share string by hashconsing the string objects (reduces memory
consumption by half on the Gcd example of certicoq). We do it for whole
idents right now, the current Coq string representation does not allow
to share prefixes sadly.

Also remove the externalisation + elaboration phases when introducing
quoted terms, we can bypass it completely and give the term directly for
the kernel to check (without bypassing any checks though!).
There is still an inefficiency due to retypechecking strings which does
not seem so easy to avoid.
We went from 240s to 24s when quoting CertiCoq's Gcd example.