[typing] add red_cast rule to avoid problems with casts in eq_term
In the cleaned-up pcuic formalization, this can be implemented by a
dedicated [cast : forall (A : Type), A -> A] constant with the same
reductions. Now we can freely use the [leq_term] to compare normal
forms: they can be made cast-free by reductions.

Variants eq/leq_term_nocast first do stripping of casts and can be
used to compare terms with the same results as the implementation of
term comparisions in Coq's kernel.