Phase split the InType TemplateMonad.
the idea of this PR is to phase split the TemplateMonad type so that
it can be extracted to Ocaml and run natively.

there are two main changes:
1. we drop TmQuote and TmQuoteRec
2. constructors which used to quote Gallina terms now take `Ast.term`s

to construct these `Ast.term`s, we use quoting `<% f x %>` quotes the
term `f x` **at elaboration time**. Some examples,

- <% 1 + 1 %> ~ (tApp <% plus %> [<% 1 %> , <% 1 %>])

free variables in the term (even if they are bound outside the term
are quoted as variables. For example,

- <% f x %> ~ (tApp (tVar "f") (tVar "x"))

this is unfortunate, but the best we can do with the simple quoting
mechanism that we have. i took a brief look at how to code splicing
but i'm not sure how to implement this.