an attempt at implementing the Extractable monad.
this isn't complete, but I think it is a step in the right direction.
i changed the following (which maybe i shouldn't have since i don't
recall what prompted the design decision).
- Monad operations that supported `option reductionStrategy` are not
  *defined* in terms of a primitive that doesn't support reduction and
   `tmEval`. This includes `tmDefinition`, `tmLemma`, `tmAxiom`, and
   `tmInferInstance`. This change is only in the `Extractable` monad,
   but I think the same should be done in the normal monad as well.

i think that a good implementation strategy would be to write normal
Ocaml functions that implement the primitives in the monad and set up
extraction to link only against that file. the interpreter will still
need the rest of the dependencies to handle quoting and anti-quoting.