Show that lift/subst can also be implemented using renamings and instantiations.
Idea suggested by Y. Forster. This brings \sigma-calculus primitives in,
with which we can hopefully prove the lifting property of the rho
function in PCUICConfluence more easily. Due to the `nat -> nat`/`nat -> term`
encoding of renamings/substitutions in this presentation, we will
however need to prove boring extensionality lemmas.