[coq] Adapt to coq/coq#11836
As parts of Coq are modified into a functional style of state passing,
we modify the MetaCoq monad to take this into account by making it
into a state-passing monad.

This is in particular needed as MetaCoq will add obligations; note
that this is for example not well defined upstream when used inside a
proof, we note so in the comments.