Finish tApp case and work on fallback case
For the tApp case (i.e. isconv_args) we have to refactor slightly.
Previously isconv_args was being passed a proof that the heads are
convertible and then produced a proof that the entire mkApps will be
convertible. The opposite is not true: if the heads are convertible and
the args are not convertible, it is still possible for the full mkApps
to be convertible (eg. constants with unused args).

This also uncovered a source of incompleteness for tCase and tFix where
safe conversion rejected two tCase/two tFix that were syntactically
equal if their args weren't convertible.  Fix this as part of this
change as well.

To remedy this change isconv_args to return a proof that the arguments
are (or aren't) convertible pairwise and then move the handling up to
the places where we are using isconv_args.

Also start working on the fallback case. It now takes another argument
putting a constraint on the shape of the heads. Refactor some universe
stuff to make it a little more uniform (in particular, no need to case
on whether we are doing Conv/Cumul) anymore.

Other than that, some slight refactorings.