Add missing reduction for predicate types
The safe checker was destructing the inferred type of predicates
directly as an arity, which was not complete. It meant that the
following example was not accepted by the safe checker:

Definition WrappedType := Type.
Definition WrappedNat : WrappedType := nat.
Definition foo : nat :=
  match 0 return WrappedNat with
  | 0 => 0
  | S n => 0
  end.

To fix we now use reduction for the predicate type.