Turn `normalisation` into a typeclass
We use two typeclasses: one for specifying that we can achieve
normalisation within a particular environment, while the other posits
normalisation for all well-formed environments.  This is useful because
while the latter is expected to be unprovable, it may be possible to
prove that specific concrete environments are normalising, which may
allow us to prove particular programs well-typed more easily without
axioms.