Allow environment weakening to reorder declarations
This is a significant reorganization of the handling of declaration
extension, with the goal of allowing environment weakening across
reordering of independent declarations.  This will be useful for using
the safechecker to concretely prove different constructions well-typed
in minimal global environments, and then combine the well-typedness
proofs in a union of the global environments without having to care
about strict extension.

We define four notions of environment extension.  The two configurable
bits are: is universe and retroknowledge extension strict (`Logic.eq`)
or loose (`⊂_cs` / `Retroknowledge.extends`); and is declaration
extension strict (fully order-preserving, i.e., new declarations are
added only at the front) or lax (declarations may be reordered and added
freely, as long as all new declarations with existing names come before
the old ones with the same names, and the relative order of declarations
with identical names is preserved).

In most cases, we are actually interested only in duplicate-free
environments, where the lax construction is equivalent to merely
requiring that the lookup function agrees on all existing delcarations.
However, we formulate the property in a way that makes sense for
duplicative environments so that strict declaration extension will
always imply lax declaration extension.  (The lookup function prefers
earlier / newer declarations over older ones.)

We thus have the following implication structure:
```
┌-----------------------------┬------------------------┬---------------------------┐
|    univ/retro extension is: |        strict          |  lax                      |
├ declaration exension is ----┼------------------------┼---------------------------┤
|    lax                      |     extends_decls      →         extends           |
|                             |            ↑           ↗            ↑              |
|   strict                    | strictly_extends_decls → extends_strictly_on_decls |
└-----------------------------┴------------------------┴---------------------------┘
```

In general, theorems use the strictest version permissible in the
conclusion (more generally in covariant positions), and the laxest
version permissible in their hypotheses (more generally in contravariant
positions).  Some theorems are parameterized over their notion of
context extension, when the version in the conclusion and the hypotheses
(more generally: isovariant positions) must match.

The exception to this rule is `SingletonProp`, `Computational`,
`Informative`, `erases`, and their dependents.  These definitions and
theorems uniformly use `strictly_extends_decls`, because I wasn't sure
if it was worth the effort to generalize these definitions over their
notion of extension, and the definitions seem to be used in both co- and
contra-variant positions.