Prove an `iff` spec for `consistent_extension_on`
Because the spec connecting
`uGraph.wGraph.IsFullSubgraph.is_full_extension` to
`consistent_extension_on` only works when the graphs satisfy the global
invariants (such as containing `lzero`) and when the first graph is a
subgraph of the second, we route through an alternate construction of
the arguments to `consistent_extension_on` where we add refresh missing
levels and then add them to the declared levels.  This is a pain, but it
works.